+++ simulation/src/nest_simulation/ant/nest_expansion.rs
@@ -85,7 +85,7 @@ pub fn ants_nest_expansion(
             commands.spawn_pheromone(
                 dirt_position,
                 Pheromone::Tunnel,
-                PheromoneStrength::new(settings.tunnel_length, settings.tunnel_length),
+                PheromoneStrength::new(settings.tunnel_length as f32, settings.tunnel_length as f32),
                 AtNest,
             );
         }
